+++ src/pages/Invite/index.tsx
@@ -98,7 +98,7 @@ const Invite: React.FC = () => {
         if (res.status === 200) {
           const data = res.data as ParticipantData;
 
-          analytics.logEvent("join_group", {
+          analytics().logEvent("join_group", {
             method: "invite",
             group_id: data.group_id,
           });